requirements.txt
alembic==1.2.1
cffi==1.15.1
chardet==3.0.4
click==8.1.3
cryptography==39.0.0
Flask==2.2.2
Flask-Cors==3.0.8
Flask-Jsonpify==1.5.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==3.0.2
idna==2.8
importlib-metadata==5.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.2
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2022.6
requests==2.22.0
SQLAlchemy==1.4.46
SQLAlchemy-Utils==0.39.0
tqdm==4.38.0
urllib3==1.25.6
yarl==1.8.1
zipp==3.11.0
psycopg2==2.9.5